requirements.txt
fastapi
uvicorn
uvloop
pandas
matplotlib
numpy
seaborn
scikit-learn
requests
python-dotenv